ChooseFilePage: moved to utils
This QWizardPage is common to Station and Rollingstock import
