+++ src/rollingstock/importer/pages/CMakeLists.txt
@@ -1,12 +1,10 @@
 set(MR_TIMETABLE_PLANNER_SOURCES
   ${MR_TIMETABLE_PLANNER_SOURCES}
-  rollingstock/importer/pages/choosefilepage.h
   rollingstock/importer/pages/fixduplicatesdlg.h
   rollingstock/importer/pages/itemselectionpage.h
   rollingstock/importer/pages/loadingpage.h
   rollingstock/importer/pages/optionspage.h
 
-  rollingstock/importer/pages/choosefilepage.cpp
   rollingstock/importer/pages/fixduplicatesdlg.cpp
   rollingstock/importer/pages/itemselectionpage.cpp
   rollingstock/importer/pages/loadingpage.cpp
