+++ src/utils/wizard/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(MR_TIMETABLE_PLANNER_SOURCES
+  ${MR_TIMETABLE_PLANNER_SOURCES}
+  utils/wizard/choosefilepage.h
+  utils/wizard/choosefilepage.cpp
+  PARENT_SCOPE
+)
